worker/requirements.txt
boto3==1.35.0
Pillow==11.0.0
